+++ pykis/responses/__init__.py
@@ -145,4 +145,4 @@ __all__ = [
     "OrderResvCcnlItem",
 ]
 
-__version__ = "1.3.0"
+__version__ = "1.3.2"